creates working smiths page
simply navigating to the smiths landing page within the
harvester file. the Smiths class inherits from the
Store class. the Harvester class is refined to have two
getDriver methods, one is more simple the other calls
for a download directory and asks to specify the mimetype
of a wanted download.